FEAT[#10] 컨트롤러 성공응답 코드 수정 및 FoodItem, Nutrients를 공통 DTO로 분리
