+++ src/main/java/MeowMeowPunch/pickeat/domain/diet/controller/FoodQueryController.java
@@ -1,7 +1,6 @@
 package MeowMeowPunch.pickeat.domain.diet.controller;
 
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -19,15 +18,12 @@ public class FoodQueryController {
 	private final FoodQueryService foodQueryService;
 
 	@GetMapping("/list/search")
-	public ResponseEntity<ResTemplate<FoodListResponse>> getFoodList(
+	public ResTemplate<FoodListResponse> getFoodList(
 		// @RequestHeader(name = "Authorization", required = false) String authorization,
 		@RequestParam(name = "cursor", required = false) String cursor,
 		@RequestParam(name = "size", required = false) Integer size
 	) {
 		FoodListResponse data = foodQueryService.getFoodList(cursor, size);
-
-		return ResponseEntity.ok(
-			new ResTemplate<>(HttpStatus.OK, "음식 리스트 조회 성공", data)
-		);
+		return new ResTemplate<>(HttpStatus.OK, "음식 리스트 조회 성공", data);
 	}
 }
